+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/ui/common/quest/QuestHeader.kt
@@ -1,8 +1,7 @@
-package de.westnordost.streetcomplete.quests
+package de.westnordost.streetcomplete.ui.common.quest
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -27,11 +26,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
 import de.westnordost.streetcomplete.ui.common.InfoFilledIcon
 import de.westnordost.streetcomplete.ui.common.InfoOutlineIcon
 import de.westnordost.streetcomplete.ui.ktx.fadingHorizontalScrollEdges
